Add separate test for float compatibility and create tests for createSimulation method.

+++ src/main/java/info/trekto/jos/core/numbers/NumberFactory.java
@@ -22,6 +22,6 @@ public interface NumberFactory {
     Number createPi();
 
     enum NumberType {
-        DOUBLE, FLOAT, ARBITRARY_PRECISION
+        FLOAT, DOUBLE, ARBITRARY_PRECISION
     }
 }
